fix: Resolver bugs de retorno e adequar a mudança da entidade

+++ src/controllers/subjectsController.js
@@ -4,11 +4,15 @@ import service from '../services/subjectsService.js';
 const routes = express.Router();
 
 routes.post('/', async (request, response) => {
-    const { subject_name } = request.body;
+    const { subject_name, course_id } = request.body;
 
     try {
         if (!subject_name || subject_name.length < 1) {
             return response.status(400).send({ message: 'A matéria precisa de um nome!' });
+        }
+
+        if (!course_id || isNaN(course_id)) {
+            return response.status(400).send({ message: 'A matéria precisa de um curso válido!' });
         }
 
         if (subject_name.length > 100) {
@@ -20,7 +24,7 @@ routes.post('/', async (request, response) => {
             return response.status(409).send({ message: 'Uma matéria com esse nome já existe!' });
         }
 
-        await service.createSubjects(subject_name);
+        await service.createSubjects(subject_name, course_id);
 
         return response.status(201).send({
             message: 'Matéria cadastrada com sucesso!',
@@ -45,6 +49,11 @@ routes.delete('/:idSubject', async (request, response) => {
             });
         }
 
+        const subject = await service.getSubject(idSubject);
+        if(!subject){
+            return response.status(400).send({ message: "Matéria não existe" })
+        }
+
         const result = await service.deleteSubject(idSubject);
 
         if (result.affectedRows === 0) {
@@ -66,7 +75,7 @@ routes.delete('/:idSubject', async (request, response) => {
 
 routes.put('/:id', async (request, response) => {
     const { id } = request.params;
-    const { subject_name } = request.body;
+    const { subject_name, course_id } = request.body;
 
     try {
         if (!id || isNaN(id)) {
@@ -75,13 +84,12 @@ routes.put('/:id', async (request, response) => {
             });
         }
 
-        if (!subject_name ) {
-            return response.status(400).send({
-                message: "Nome da matéria não fornecido ou inválido."
-            });
+        const existingSubject = await service.findSubjectByName(subject_name);
+        if (existingSubject) {
+            return response.status(409).send({ message: 'Uma matéria com esse nome já existe!' });
         }
 
-        const result = await service.updateSubject(id, subject_name);
+        const result = await service.updateSubject(id, subject_name, course_id);
 
         if (result.affectedRows === 0) {
             return response.status(404).send({
@@ -94,7 +102,7 @@ routes.put('/:id', async (request, response) => {
         });
     } catch (error) {
         return response.status(500).send({
-            message: "Erro ao atualizar a matéria.",
+            message: "Erro ao atualizar os dados da matéria.",
             error: error.message
         });
     }
